修改sortVO+commodityVO

+++ JXCSystemClient/src/main/java/vo/VoucherVO.java
@@ -4,12 +4,12 @@ public class VoucherVO {
 	
 	public String start_time;
 	public String end_time;
-	public int start_money;
-	public int end_money;
-	public int money;
+	public double start_money;
+	public double end_money;
+	public double money;
 	public int level;
 	
-	public VoucherVO(String start_time,String end_time,int start_money,int end_money,int money,int level){
+	public VoucherVO(String start_time,String end_time,double start_money,double end_money,double money,int level){
 		this.start_time=start_time;
 		this.end_time=end_time;
 		this.start_money=start_money;
@@ -25,13 +25,13 @@ public class VoucherVO {
 		return end_time;
 	}
 	
-	public int getStartMoney(){
+	public double getStartMoney(){
 		return start_money;
 	}
-	public int getEndMoney(){
+	public double getEndMoney(){
 		return end_money;
 	}
-	public int getMoney(){
+	public double getMoney(){
 		return money;
 	}
 	public int getLevel(){
